+++ src/tree_functions.cpp
@@ -5,34 +5,33 @@
 #include "counter.hpp"
 #include "utils.hpp"
 
-double entropy(const Counter& counters)
+float entropy(const Counter& counters)
 {
     size_t size = 0;
     for (size_t i = 0; i < counters.size(); ++i)
         size += counters[i];
 
-    double e = 0.0;
-    double proportion;
+    float e = 0.0;
+    float proportion;
     for (size_t i = 0; i < counters.size(); ++i)
     {
         if (counters[i] == 0)
             continue;
 
-        proportion = (double)counters[i] / size;
+        proportion = (float)counters[i] / size;
         e -= proportion * std::log2(proportion);
     }
 
     return e;
 }
 
-double entropy(const std::vector<uint32_t>& y,
-               const std::vector<size_t>& indices)
+float entropy(const std::vector<uint8_t>& y, const std::vector<size_t>& indices)
 {
     return entropy(count(y, indices));
 }
 
-double informationGain(double parent_entropy, const Counter& left,
-                       const Counter& right)
+float informationGain(float parent_entropy, const Counter& left,
+                      const Counter& right)
 {
     size_t left_size = 0;
     for (size_t i = 0; i < left.size(); ++i)
@@ -47,11 +46,11 @@ double informationGain(double parent_entropy, const Counter& left,
 
     size_t size = left_size + right_size;
 
-    double e_left = entropy(left);
-    double e_right = entropy(right);
+    float e_left = entropy(left);
+    float e_right = entropy(right);
 
-    double ratio_left = (double)left_size / size;
-    double ratio_right = (double)right_size / size;
+    float ratio_left = (float)left_size / size;
+    float ratio_right = (float)right_size / size;
 
     return parent_entropy - (ratio_left * e_left + ratio_right * e_right);
 }
